+++ .storybook/preview-head.html
@@ -1,5 +1,5 @@
 <script>
-    // https://github.com/storybookjs/storybook/issues/18399
-    // Fix: add `jest` to global
-    window.global = window;
+  // https://github.com/storybookjs/storybook/issues/18399
+  // Fix: add `jest` to global
+  window.global = window;
 </script>
