--- src/components/Error/Error.tsx
@@ -1,9 +1,0 @@
-import "@/index.css";
-
-export type ErrorProps = {
-  string: string;
-};
-
-export const Error = (p: ErrorProps): JSX.Element => (
-  <p className="mx-4">Error: {p.string}</p>
-);
