+++ src/components/FloatingToolbar/FloatingToolbar.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 
-import { FloatingToolbar, FloatingToolbarProps } from "./FloatingToolbar";
+import { FloatingToolbar, FloatingToolbarProps } from "./";
 
 export default {
   title: "Application/Component Library/FloatingToolbar",
